Start adding bash completion
Add ability to generate most basic bash completion. `make completion`
prints a completion shell script to stdout.
