+++ cmd/octopus/bashcompletion/generate.go
@@ -1,0 +1,12 @@
+package main
+
+import (
+	"os"
+
+	"github.com/BlaineEXE/octopus/cmd/octopus/config"
+	_ "github.com/BlaineEXE/octopus/cmd/octopus/root"
+)
+
+func main() {
+	config.OctopusCmd.GenBashCompletion(os.Stdout)
+}
