+++ cmd/octopus/main.go
@@ -9,19 +9,8 @@ import (
 	"log"
 
 	"github.com/BlaineEXE/octopus/cmd/octopus/config"
-	"github.com/BlaineEXE/octopus/cmd/octopus/copy"
-	"github.com/BlaineEXE/octopus/cmd/octopus/run"
-	"github.com/BlaineEXE/octopus/cmd/octopus/version"
+	_ "github.com/BlaineEXE/octopus/cmd/octopus/root"
 )
-
-func init() {
-	octopusCmd := config.OctopusCmd
-
-	// Subcommands
-	octopusCmd.AddCommand(version.VersionCmd)
-	octopusCmd.AddCommand(run.RunCmd)
-	octopusCmd.AddCommand(copy.CopyCmd)
-}
 
 func main() {
 	if err := config.OctopusCmd.Execute(); err != nil {
